PinLock v2: Proposal PIN unlock interface in Gaia
With PIN handling landing in Gecko code, we need a Gaia interface to ask
user for its pin code and communicate with the Gecko interfaces to
effectively perform the unlocking. This is a tentative proposal that
should be working more or less: at least it works on my side (Nexus S,
Free Mobile).

As it is a tentative, it lacks:
 - PUK handling
 - PIN code changing (I don't think it's event the right place for)

There is few feedback: the keypad should disappear once the pin is
entered, validated and accepted by the system. It means if you give it a
bad pin code, it will keep asking. Moreover, it does not indicate how
far in retry you are. A nice enhancement would be to notify the user of
the unlocking process (starting, running, finished successfully or
error) before just disappearing.