Fixing orientation changes on Nexus S
Re-using an old hack, it seems that the hack needs to be in both file in
order to have an effect now. No idea why. Tried only in bootstrap.js, or
only in window_manager.js, no success. Device orientation handling only
got working back once the hack was in both.