Bug 802598/802603 - [Apps] App Download Progress in Notification Center r=etienne
This patch adds a new dom element in the notification center
to be able to adds advanced notifications that the normal
notification center can't do. We do one notification per installed app.

It tries to put some things in common with the update-manager.

It also changes the way these fake notifications contribute to the
existing notification center (NotificationScreen object), and some tests
were written or fixed.

It steals the progress bar that was already used in the popup manager,
and it's also the same that is used in the browser app.

It lays out a (hopefully) better way to do mocks.