Bug 815203 - [system] handle the app install restart when the homescreen starts it r=etienne r=arcturus
We used to display the notification as soon as we got the applicationinstall
event. Now we display the notification in the notification center only when
the first progress event is fired on the app object.

When an app install is restarted, we'll get only progress events, then it
fits correctly in our new way of displaying the notification.

We take also great care to remove the handlers on the app object only if
they're our own handlers.

We also have to add the handlers on pending apps during init to listen for
app install restarts.