Bug 1173782 - Close LogShake notification only on MozActivity success r=daleharvey
For the sake of being user friendly, we want to keep logshake
notifications until they have been really used. In the past we dropped
them as soon as the user did click on it. This commit will make sure we
only call close code if the triggered MozActivity did succeeded.